btrfs-progs: tests: add runtime check for free-space-tree
The CI hosts have old kernel that does not support the FST, make the
test fail gracefully.

+++ tests/fsck-tests/037-freespacetree-repair/test.sh
@@ -50,6 +50,11 @@ corrupt_fst_item()
 		-f offset "$TEST_DEV"
 }
 
+if ! [ -f "/sys/fs/btrfs/features/free_space_tree" ]; then
+	_not_run "kernel does not support free-space-tree feature"
+	exit
+fi
+
 run_check "$TOP/mkfs.btrfs" -n 4k -f "$TEST_DEV"
 run_check_mount_test_dev -oclear_cache,space_cache=v2
